requirements.txt
dj-database-url==0.5.0
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
gunicorn==20.1.0
psycopg2-binary==2.9.3
wheel==0.37.1
whitenoise==6.0.0